pyproject.toml
[project]
name = "openclaw-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "httpx[socks]>=0.28.1",
    "openai>=1.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
]
